+++ frontend/src/routes/_oh._index/route.tsx
@@ -5,9 +5,9 @@ import { setReplayJson } from "#/state/initial-query-slice";
 import { useGitUser } from "#/hooks/query/use-github-user";
 import { useGitHubAuthUrl } from "#/hooks/use-github-auth-url";
 import { useConfig } from "#/hooks/query/use-config";
-import { GitRepositoriesSuggestionBox } from "#/components/features/github/git-repositories-suggestion-box";
+import { GitRepositoriesSuggestionBox } from "#/components/features/git/git-repositories-suggestion-box";
 import { ReplaySuggestionBox } from "../../components/features/suggestions/replay-suggestion-box";
-import { CodeNotInGitHubLink } from "#/components/features/github/code-not-in-github-link";
+import { CodeNotInGitHubLink } from "#/components/features/git/code-not-in-github-link";
 import { HeroHeading } from "#/components/shared/hero-heading";
 import { TaskForm } from "#/components/shared/task-form";
 import { convertFileToText } from "#/utils/convert-file-to-text";
